28 feb

+++ src/main/java/com/example/demo/student/StudentService.java
@@ -1,5 +1,7 @@
 package com.example.demo.student;
 
+import com.example.demo.student.exception.BadRequestException;
+import com.example.demo.student.exception.StudentNotFoundException;
 import java.util.List;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -15,10 +17,17 @@ public class StudentService {
    }
 
    public void addStudent(Student student) {
+      boolean existsEmail = studentRepository.selectExistsEmail(student.getEmail());
+      if (existsEmail) {
+         throw new BadRequestException("Email " + student.getEmail() + " taken");
+      }
       studentRepository.save(student);
    }
 
    public void deleteStudent(Long studentId) {
+      if (!studentRepository.existsById(studentId)) {
+         throw new StudentNotFoundException("Student with ID " + studentId + "does not exist");
+      }
       studentRepository.deleteById(studentId);
    }
 }
